refactor: mv multicall updater to lib
