--- src/lib/state/multicall.ts
@@ -1,8 +1,0 @@
-import { createMulticall } from '@uniswap/redux-multicall'
-import { combineReducers, createStore } from 'redux'
-
-export const multicall = createMulticall()
-const reducer = combineReducers({ [multicall.reducerPath]: multicall.reducer })
-export const store = createStore(reducer)
-
-export default multicall
