--- src/state/multicall/updater.tsx
@@ -1,13 +1,0 @@
-import useBlockNumber from 'lib/hooks/useBlockNumber'
-import multicall from 'lib/state/multicall'
-
-import { useInterfaceMulticall } from '../../hooks/useContract'
-import { useActiveWeb3React } from '../../hooks/web3'
-
-// Create Updater wrappers that pull needed info from store
-export default function Updater() {
-  const latestBlockNumber = useBlockNumber()
-  const { chainId } = useActiveWeb3React()
-  const multicall2Contract = useInterfaceMulticall()
-  return <multicall.Updater chainId={chainId} latestBlockNumber={latestBlockNumber} contract={multicall2Contract} />
-}
